+++ src/greynoise/cli/formatter.py
@@ -5,14 +5,13 @@ from __future__ import print_function
 
 import functools
 import json
-from xml.dom.minidom import parseString
 
 import ansimarkup
 import click
 import colorama
-from jinja2 import Environment, PackageLoader, select_autoescape
 
-from dicttoxml import dicttoxml
+from dict2xml import dict2xml
+from jinja2 import Environment, PackageLoader, select_autoescape
 
 JINJA2_ENV = Environment(
     loader=PackageLoader("greynoise.cli"),
@@ -59,7 +58,15 @@ def json_formatter(result, _verbose):
 
 def xml_formatter(result, _verbose):
     """Format result as xml."""
-    return parseString(dicttoxml(result)).toprettyxml()
+    xml_formatted = ""
+    if type(result) is list:
+        xml_formatted = dict2xml({"item": result}, wrap="root", indent="\t")
+    else:
+        xml_formatted = dict2xml(result, wrap="root", indent="   ")
+
+    # dict2xml does not add header, so add header manually
+    xml_header = '<?xml version="1.0" ?>'
+    return "{}\n{}".format(xml_header, xml_formatted)
 
 
 def get_location(metadata):