+++ tests/test_api.py
@@ -32,12 +32,14 @@ class TestInit(object):
             "api_key": "<api_key>",
             "api_server": "<api_server>",
             "timeout": "<timeout>",
+            "proxy": "<proxy>",
         }
         with patch("greynoise.api.load_config") as load_config:
             client = GreyNoise(**config)
             assert client.api_key == config["api_key"]
             assert client.api_server == config["api_server"]
             assert client.timeout == config["timeout"]
+            assert client.proxy == config["proxy"]
             load_config.assert_not_called()
 
     def test_without_api_key(self):
@@ -46,6 +48,7 @@ class TestInit(object):
             "api_key": "<api_key>",
             "api_server": "<api_server>",
             "timeout": "<timeout>",
+            "proxy": "<proxy>",
         }
         with patch("greynoise.api.load_config") as load_config:
             load_config.return_value = config
@@ -53,6 +56,7 @@ class TestInit(object):
             assert client.api_key == config["api_key"]
             assert client.api_server == config["api_server"]
             assert client.timeout == config["timeout"]
+            assert client.proxy == config["proxy"]
             load_config.assert_called()
 
 
@@ -302,7 +306,10 @@ class TestFilter(object):
                 "0.0.0.0\n255.255.255.255\nnot an ip address",
                 "<noise>0.0.0.0</noise>\n",
             ),
-            ("0.0.0.0 255.255.255.255\nnot an ip address", "",),
+            (
+                "0.0.0.0 255.255.255.255\nnot an ip address",
+                "",
+            ),
         ],
     )
     def test_select_noise(self, client, text, expected_output):
@@ -639,4 +646,17 @@ class TestStats(object):
         client._request.assert_called_with(
             "experimental/gnql/stats", params={"query": query}
         )
+        assert response == expected_response
+
+
+class TestMeta(object):
+    """GreyNoise client run GNQL stats query test cases."""
+
+    def test_metadata(self, client):
+        """Run GNQL stats query."""
+        expected_response = []
+
+        client._request = Mock(return_value=expected_response)
+        response = client.metadata()
+        client._request.assert_called_with("meta/metadata")
         assert response == expected_response