+++ tests/test_util.py
@@ -22,6 +22,7 @@ class TestLoadConfig(object):
             "api_key": "",
             "api_server": "https://api.greynoise.io",
             "timeout": 60,
+            "proxy": "",
         }
 
     @patch("greynoise.util.open")
@@ -32,6 +33,7 @@ class TestLoadConfig(object):
             "api_key": "<api_key>",
             "api_server": "<api_server",
             "timeout": 123456,
+            "proxy": "",
         }
 
         os.environ = {}
@@ -42,8 +44,12 @@ class TestLoadConfig(object):
             api_key = {}
             api_server = {}
             timeout = {}
+            proxy = {}
             """.format(
-                expected["api_key"], expected["api_server"], expected["timeout"],
+                expected["api_key"],
+                expected["api_server"],
+                expected["timeout"],
+                expected["proxy"],
             )
         )
         open().__enter__.return_value = StringIO(file_content)
@@ -60,6 +66,7 @@ class TestLoadConfig(object):
             "api_key": "<api_key>",
             "api_server": "<api_server>",
             "timeout": 123456,
+            "proxy": "",
         }
 
         os.environ = {"GREYNOISE_API_KEY": expected["api_key"]}
@@ -70,8 +77,11 @@ class TestLoadConfig(object):
             api_key = unexpected
             api_server = {}
             timeout = {}
+            proxy = {}
             """.format(
-                expected["api_server"], expected["timeout"],
+                expected["api_server"],
+                expected["timeout"],
+                expected["proxy"],
             )
         )
         open().__enter__.return_value = StringIO(file_content)
@@ -88,6 +98,7 @@ class TestLoadConfig(object):
             "api_key": "<api_key>",
             "api_server": "<api_server>",
             "timeout": 123456,
+            "proxy": "",
         }
 
         os.environ = {"GREYNOISE_API_SERVER": expected["api_server"]}
@@ -98,8 +109,11 @@ class TestLoadConfig(object):
             api_key = {}
             api_server = unexpected
             timeout = {}
+            proxy = {}
             """.format(
-                expected["api_key"], expected["timeout"],
+                expected["api_key"],
+                expected["timeout"],
+                expected["proxy"],
             )
         )
         open().__enter__.return_value = StringIO(file_content)
@@ -116,6 +130,7 @@ class TestLoadConfig(object):
             "api_key": "<api_key>",
             "api_server": "<api_server>",
             "timeout": 123456,
+            "proxy": "",
         }
 
         os.environ = {"GREYNOISE_TIMEOUT": str(expected["timeout"])}
@@ -126,8 +141,11 @@ class TestLoadConfig(object):
             api_key = {}
             api_server = {}
             timeout = unexpected
+            proxy = {}
             """.format(
-                expected["api_key"], expected["api_server"],
+                expected["api_key"],
+                expected["api_server"],
+                expected["proxy"],
             )
         )
         open().__enter__.return_value = StringIO(file_content)
@@ -144,6 +162,7 @@ class TestLoadConfig(object):
             "api_key": "<api_key>",
             "api_server": "<api_server>",
             "timeout": 123456,
+            "proxy": "",
         }
 
         os.environ = {"GREYNOISE_TIMEOUT": "invalid"}
@@ -154,8 +173,12 @@ class TestLoadConfig(object):
             api_key = {}
             api_server = {}
             timeout = {}
+            proxy = {}
             """.format(
-                expected["api_key"], expected["api_server"], expected["timeout"],
+                expected["api_key"],
+                expected["api_server"],
+                expected["timeout"],
+                expected["proxy"],
             )
         )
         open().__enter__.return_value = StringIO(file_content)
@@ -174,6 +197,7 @@ class TestSaveConfig(object):
             "api_key": "<api_key>",
             "api_server": "<api_server>",
             "timeout": 123456,
+            "proxy": "",
         }
 
         with patch("greynoise.util.os") as os, patch("greynoise.util.open") as open_:
@@ -190,6 +214,7 @@ class TestSaveConfig(object):
             "api_key": "<api_key>",
             "api_server": "<api_server>",
             "timeout": 123456,
+            "proxy": "",
         }
         expected = textwrap.dedent(
             """\
@@ -197,9 +222,12 @@ class TestSaveConfig(object):
             api_key = {}
             api_server = {}
             timeout = {}
-
+            proxy = {}\n
             """.format(
-                config["api_key"], config["api_server"], config["timeout"],
+                config["api_key"],
+                config["api_server"],
+                config["timeout"],
+                config["proxy"],
             )
         )